Port the system to the Pi Zero W, and fix what 32 bits exposed
The board that arrived is a Zero W, not a Zero 2 W: an ARM1176, ARMv6,
32-bit. Not a slower Zero 2 W - a different instruction set, on which
nothing built for the other board executes a single instruction.

Building for it turned up six bugs that were invisible on the two
64-bit machines, because on those a long is eight bytes and every
struct the kernel writes has the shape this code assumed:

  date hung. clock_gettime was the 32-bit call at 263, handed this
  libc's 16-byte timespec: the seconds came back with the nanoseconds
  in their top half, the clock read as hundreds of billions of years
  from now, and the calendar loop never finished. Now the _time64 pair.

  logd sat at 100% of the one core. ppoll at 336 wants two 32-bit
  words; it read the seconds and took zero for the nanoseconds, so
  every wait under a second was a wait of no time at all. Now
  ppoll_time64.

  df printed 17650703386236652 1K-blocks and guard reported "/data has
  -327968716MB left". statfs on a 32-bit machine fills a struct whose
  every field is 32 bits; statfs64 is the one with 64-bit counts, and
  on ARM it is packed to four bytes. Neither number looked like a type
  error - they looked like a broken filesystem.

  wc -c said 1GB for a 5GB file, ls -l agreed, and dd skip=3000 read
  from the start. lseek at 19 takes a 32-bit offset: correct up to two
  gigabytes and then quietly not. Now _llseek, and lp_lseek returns
  s64 because a long here cannot hold the answer.

  expr 2147483647 + 1 printed -2147483648.

  The card carried an aarch64 e2fsck. It copied fine and looked right;
  the only way to find out was to damage /data and watch the repair
  fail at exec. mksdcard now checks the architecture of what it ships,
  which would also have caught the same thing in the amd64 image.

dropbear and wpa_supplicant needed a libc of their own. Debian's armhf
port has an ARMv7 baseline - its glibc uses movw/movt and its libgcc is
Thumb-2, neither of which an ARM1176 has - so build-thirdparty.sh now
builds musl for ARMv6 and links them against armel, whose libgcc is ARM
mode plus Thumb-1. The check that they run is not a file-format check:
each one is executed under qemu -cpu arm1176, which raises SIGILL for
anything the real chip lacks.

The kernel is bcmrpi_defconfig plus the arm64 fragment minus everything
that assumes a 64-bit or virtual machine, generated the same way the
amd64 config is. BCM2835_POWER is off: the device tree names the
firmware power driver, and the ASB one took an external abort inside
its probe with interrupts disabled, ending the boot at two seconds.

Also here, and not ARM-specific: df resolves a relative path (df . said
"not on any mounted filesystem", which is a strange thing to be told
about the directory you are standing in), and od seeks rather than
reading through the bytes -j skips.

Verified: 51/51 against GNU coreutils and 21/21 on the syscall layer,
on both armv6 and amd64; 149 binaries run under an ARM1176 model with
no SIGILL and no hang; and the whole system boots in QEMU raspi0 to a
shell, with SSH listening and /data readable and writable.

+++ userland/defend/defend.c
@@ -752,7 +752,8 @@ static bool save_bans(void)
     for (int i = 0; i < nbans; i++) {
         char host[20];
         ipv4_format(bans[i].addr, host);
-        dprintf((int)fd, "%s %ld %d\n", host, bans[i].when, bans[i].fails);
+        dprintf((int)fd, "%s %lld %d\n", host,
+                (long long)bans[i].when, bans[i].fails);
     }
     lp_close((int)fd);
     if (lp_rename(F_BANS ".new", F_BANS) != 0) {
@@ -1980,8 +1981,9 @@ static void save_state(void)
     long fd = lp_open(F_STATE, O_WRONLY | O_CREAT | O_TRUNC, 0600);
     if (fd < 0)
         return;
-    dprintf((int)fd, "checks %ld\nlast %ld\nbanned %ld\n",
-            st_checks, st_last, st_banned_total);
+    dprintf((int)fd, "checks %lld\nlast %lld\nbanned %lld\n",
+            (long long)st_checks, (long long)st_last,
+            (long long)st_banned_total);
     lp_close((int)fd);
 }
 
@@ -2058,7 +2060,7 @@ static int cmd_status(void)
     load_baseline();
 
     when_str(st_last, buf, sizeof buf);
-    printf("checks run     %ld, last at %s\n", st_checks, buf);
+    printf("checks run     %lld, last at %s\n", (long long)st_checks, buf);
 
     when_str(base_when, buf, sizeof buf);
     printf("baseline       %s", buf);
@@ -2068,8 +2070,8 @@ static int cmd_status(void)
         printf("  (%d items)", nbase);
     printf("\n");
 
-    printf("blocked        %d now, %ld since this file was created\n",
-           nbans, st_banned_total);
+    printf("blocked        %d now, %lld since this file was created\n",
+           nbans, (long long)st_banned_total);
 
     if (nbans > 0) {
         s64 now = lp_time();
@@ -2081,8 +2083,8 @@ static int cmd_status(void)
             when_str(bans[i].when, at, sizeof at);
             s64 left = bans[i].when + BAN_SECS - now;
             if (left < 0) left = 0;
-            printf("%-16s %-8d %-24s %ld hours\n",
-                   host, bans[i].fails, at, left / 3600);
+            printf("%-16s %-8d %-24s %lld hours\n",
+                   host, bans[i].fails, at, (long long)(left / 3600));
         }
         printf("\n`defend unban <address>` lets one back in.\n");
     }
@@ -2155,7 +2157,7 @@ static int cmd_baseline(void)
     dprintf((int)fd,
             "# What defend expects to find. Anything not listed here is\n"
             "# reported. Edit it by hand or run `defend baseline` again.\n");
-    dprintf((int)fd, "# recorded %ld\n", lp_time());
+    dprintf((int)fd, "# recorded %lld\n", (long long)lp_time());
     for (int i = 0; i < nlisten; i++)
         dprintf((int)fd, "%s\n", listen_line[i]);
     for (int i = 0; i < nkeylines; i++)
